Move lists of families to global variables in single_source
Rather than defining lists of supported filename formats and product families directly
within the code, define them as global variables at the top for clarity.

This does not seem it should have fixed anything, but it appears it may have caused
the previously failing "unsectored" test output to pass :shrug:

(previously it was attempting to calculate coverage for the "unsectored" test output,
which is not expected behaviour, then failing).

NRLMMD-GEOIPS/geoips#69