Make all order of ops only families additionalProperties: true
The product families that ONLY specify how things get called in the procflows, and
specify nothing regarding the formatting of the YAML with additionalProperties true
set on the spec in the schema.

This allows arbitrary fields in the product during development - though I think
long term specific families should be set up once the fields and format is determined.

We may also want to rename these as unvalidated_*.yaml, since the actual spec is
NOT validated at all.
NRLMMD-GEOIPS/geoips#122