Replace sector_loaders.trackfiles with sector_metadata_generators
This is in the explanation of the "trackfile_parser" command line argument.
NOTE I am not replacing the trackfile_parser argument, because that
specifically goes with the "trackfiles" argument - telling geoips
which sector_metadata_generator should specifically be used for the
trackfiles specified command line. We can revisit this later.