v1.12.0 geoips - VERSION RELEASE - plugin registry bug fixes, geoips install workflows (#419)
* 1.12.0 release NRLMMD-GEOIPS/geoips#418

* Removing unused comments

* Update installation.rst

* Update installation.rst

* Update installation.rst

* Add leading '/' to flake8 per-file-ignores

Without the leading /, flake8 would fail from certain directories.
Sometimes it would fail when running from within the repo, sometimes it would
faile when running from OUTSIDE the repo.  With the leading /, it seems to
pass from anywhere.

* Update black.yaml to use GEOIPS_ACTIVE_BRANCH

To ensure the correct version of the geoips code base, reference the GEOIPS_ACTIVE_BRANCH organization variable.  This allows updating the geoips branch to use for actions if updates are made to the workflows.

* variables -> vars

* Use v4 checkout

* Update black.yaml

* Update black.yaml

* Update black.yaml

* Update flake8.yaml for GEOIPS_ACTIVE_BRANCH

* Update black.yaml to checkout v3

* Add .config to sparse-checkout

* Add .config to sparse checkout

* Turn on build docs html

* Use org vars for runner, org

* Turn off build-html-docs

* Turn off pytest-short.yaml

* Turn off test-interfaces.yaml

* Updating black formatting from 1.12.0 release (#430)

Co-authored-by: Mindy <none>

---------

Co-authored-by: mindyls <mindyls@users.noreply.github.com>