Improve clarity about performance increase :pencil:
On PR #400 I commented that the documentation for the pluginregistry functionality needed. I outlined 3 different changes that needed to be made.

I was able to take care of concern 2:

```
The benefits section has a sentence which is difficult to understand without context. The performance increases are partly due to the new code "only instantiating the plugin registry when a user requests a plugin", but the previous behavior isn't detailed which makes it difficult to understand why (explicitly) this is an improvement.
```

1 and 3 still need take care of:

```
The example is helpful, but needs to be more explicitly explained. Eg. walk the reader through the example "This line of the file ... because .. "
```

and

```
How does the plugin registry work? We are told why it's important, that it store info and some other details. But not explicitly what it does and how.
```

:slightly_smiling_face: :pencil: