Create Documentation For PluginRegistry (#400)
* added this branch to the release notes

* 1.12.0 release NRLMMD-GEOIPS/geoips#418

* Removing unused comments

* Update installation.rst

* Update installation.rst

* Update installation.rst

* Add leading '/' to flake8 per-file-ignores

Without the leading /, flake8 would fail from certain directories.
Sometimes it would fail when running from within the repo, sometimes it would
faile when running from OUTSIDE the repo.  With the leading /, it seems to
pass from anywhere.

* moved release notes for this branch to most current version v1_12_2a0.rst

* Copyedited documentation

 - Fixed grammar and spelling
 - Copyedited for clarity, brevity and flow
 - Added comments where further clarification is needed

Related to Issue #386 and Pull request #400

* Improve clarity about performance increase :pencil:

On PR #400 I commented that the documentation for the pluginregistry functionality needed. I outlined 3 different changes that needed to be made.

I was able to take care of concern 2:

```
The benefits section has a sentence which is difficult to understand without context. The performance increases are partly due to the new code "only instantiating the plugin registry when a user requests a plugin", but the previous behavior isn't detailed which makes it difficult to understand why (explicitly) this is an improvement.
```

1 and 3 still need take care of:

```
The example is helpful, but needs to be more explicitly explained. Eg. walk the reader through the example "This line of the file ... because .. "
```

and

```
How does the plugin registry work? We are told why it's important, that it store info and some other details. But not explicitly what it does and how.
```

:slightly_smiling_face: :pencil:

* Correct doc phrasing

* Removing old human readable JSON release note

* Rearranged release note

* Fix typo, add missing word

* Update plugin_registries.rst

Replaced "encapsulate" with "represent the three available"

* Add sphinxcontrib-autoprogram

* Move parser creation to function

* Reorganize

* Add note about autobuilding registries

* Add info about sphinxcontrib-autoprogram

* Update v1_12_2a0.rst

---------

Co-authored-by: mindyls <mindyls@users.noreply.github.com>
Co-authored-by: mindyls <13873624+mindyls@users.noreply.github.com>
Co-authored-by: Gwyn Uttmark <biosafetylevel5@gmail.com>
Co-authored-by: BSL-5 <biosafetylvl5@users.noreply.github.com>
Co-authored-by: Jeremy Solbrig <1684491+jsolbrig@users.noreply.github.com>
Co-authored-by: Jeremy Solbrig <jeremy.solbrig@colostate.edu>